Added async versions of existing APIs
- Added OutputTabsAsync()
- Added WriteLineNoTabsAsync(string)
