+++ src/libraries/System.CodeDom/tests/System.CodeDom.Tests.csproj
@@ -30,7 +30,6 @@
     <Compile Include="System\CodeDom\Compiler\CompilerParametersTests.cs" />
     <Compile Include="System\CodeDom\Compiler\CompilerResultsTests.cs" />
     <Compile Include="System\CodeDom\Compiler\CSharpCodeGeneratorTests.cs" />
-    <Compile Include="System\CodeDom\Compiler\IndentedTextWriterTests.cs" />
     <Compile Include="System\CodeDom\TempFileCollectionTests.cs" />
     <Compile Include="System\CodeDom\CodeObjectCreateExpressionTests.cs" />
     <Compile Include="System\CodeDom\CodeParameterDeclarationExpressionTests.cs" />
